Add simple support for redirects
Follow 3XX status redirects with a new request. Parse the `location`
header for an absolute url, absolute path, or relative path. Replace or
amend the url from the original request and place a new request.

Convert all keys in the headers dictionary to lowercase to perform a
case-insensitive search.

There's no limit on the number of consecutive redirects. Since the
`requets()` method is now recursive, multiple redirects may crash the
stack. Especially on a CircuitPython microcontroller.

Expand the requests_simpletest_cpython.py example with new requests
utilizing the three supported types of redirects.

Note: We're using the httpbingo.org domain for redirects until
the following issue on httpbin.org is resolved:
https://github.com/postmanlabs/httpbin/issues/617

This has been tested on CPython 3.8 and an ESP32-S2 based MagTag board
running CircuitPython 6.0.