Add Rachio Irrigation API Example
Example pulls data from your Rachio Irrigation Timer. This is just a basic example. More can be done with zone scheduling, actually running the irrigation system, etc.. will work on an advanced example in the future.  This should be good enough for very basic communication with the API and timer.